Adaptações na conexão com mysql e primeira rota Novo Usuário criada

+++ BackEnd/src/infrastructure/persistence/mysql/migrations/5-accounting.migrations.mysql.db.ts
@@ -28,7 +28,7 @@ export default {
                 allowNull: false
             },
             appointments: Sequelize.DataTypes.INTEGER,
-            collected: Sequelize.DataTypes.INTEGER,
+            collected: Sequelize.DataTypes.DECIMAL(8, 2),
             createdAt: Sequelize.DataTypes.DATE,
             updatedAt: Sequelize.DataTypes.DATE,
         });
